Add book detail delete modal component  (#147)
